Refactor: Move Asstype to association.go to align with Gadgets
